requirements.txt
tweepy==4.9.0
pillow==9.1.1
